API: Moved API file directory around.
