+++ api/package.json
@@ -11,6 +11,7 @@
     "bcrypt": "^3.0.6",
     "cors": "^2.8.5",
     "express": "^4.16.4",
+    "package.json": "^2.0.1",
     "pg": "^7.10.0",
     "ts-node-dev": "^1.0.0-pre.32",
     "typescript": "^3.4.1"
